Create SmsHelperService and move helper functionality out of TwillioService and into the SmsHelperService. Also fix an error in helm-config.
