+++ src/remote/twillio.service.interface.ts
@@ -1,4 +1,3 @@
 export abstract class ITwillioService {
     abstract async sendOtp(toNumber: string, otp: number): Promise<void>;
-    abstract generateRandomOtp(): number;
 }
