+++ src/sms/sms.module.ts
@@ -4,6 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { SmsOtp } from '../entity/sms.otp';
 import { RateLimitModule } from '../ratelimit/ratelimit.module';
 import { RemoteModule } from '../remote/remote.module';
+import { SmsHelperService } from './sms.helper.service';
 
 /**
  * The SMS module doesn't expose any endpoints of it's own so doesn't need a controller
@@ -14,7 +15,7 @@ import { RemoteModule } from '../remote/remote.module';
         RateLimitModule,
         RemoteModule
     ],
-    providers: [SmsService],
+    providers: [SmsService, SmsHelperService],
     exports: [SmsService]
 })
 export class SmsModule {}
