+++ test/mock/mock.sms.helper.service.ts
@@ -1,0 +1,11 @@
+import { SmsHelperService } from '../../src/sms/sms.helper.service';
+
+export class MockSmsHelperService extends SmsHelperService {
+    constructor(private readonly otp?: number) {
+        super();
+    }
+
+    generateRandomOtp(): number {
+        return this.otp ?? super.generateRandomOtp();
+    }
+}
